requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
customtkinter==5.2.2
darkdetect==0.8.0
frozenlist==1.7.0
idna==3.10
multidict==6.6.4
packaging==24.0
pillow==11.3.0
propcache==0.3.2
pytubefix==9.5.0
typing_extensions==4.15.0
yarl==1.20.1
